added convergence criteria
Here, we can directly access the model updates from the optimiser which is equivalent to (x_old - x). This bypasses the need to look at gradients directly, which was what I was doing previously. If the L2 norm of the updates is less than the user-defined relative tolerance of the current data then we have converged.