added convergence criteria (#21)
* added convergence criteria

Here, we can directly access the model updates from the optimiser which is equivalent to (x_old - x). This bypasses the need to look at gradients directly, which was what I was doing previously. If the L2 norm of the updates is less than the user-defined relative tolerance of the current data then we have converged.

* convergence criteria tracks pyTree leaves

* forgot to turn back on JIT compilation

* single tree_map to check for convergence

* jit computation of norms for convergence test

* removed scene copy; callback gets convergence tree

---------

Co-authored-by: Peter Melchior <peter.m.melchior@gmail.com>