Fix variable type bug in Observation
I'm fairly sure this is a bug - if a `channels` list is not provided when instantiating an Observation object, a default will be created, but here we are just creating a `range(...)`. 

In the frame.py module, something similar is happening, but there is the default is `channels = list(range(data.shape[0]))`. 

Additionally creating an Observation instance will raise an error on this line because of `range`.